Separation of main and auxiliary
Added source file containing auxiliary functions for easier use and maintenance of the program
